fix and complete zigbee thread

+++ linux_zigbee_coor/rbps/rbps_glb.h
@@ -14,7 +14,8 @@
 #define RBPS_GLB_H_
 
 /* Includes */
-#include "stdint.h"
+#include <cstdint>
+#include <cstdio>
 
 /* definitions */
 #define RBPS_DEBUG (1)
@@ -54,7 +55,7 @@ namespace rbps_ns {
 	extern int wi2cc_mq_id;
 
 	/* message mtext max size */
-	const uint8_t mtext_max_size = 3+255;
+	const uint16_t mtext_max_size = 3+255;
 	/* message type */
 	typedef struct mesg_s {
 		long mtype;
